fixed integer overflow

+++ maths.cpp
@@ -1,11 +1,11 @@
 #include "maths.hpp"
 
-int add(int x, int y)
+long long add(int x, int y)
 {
   return x + y;
 }
 
-int mult(int x, int y)
+long long mult(int x, int y)
 {
   return x * y;
 }
