+++ maths.hpp
@@ -1,9 +1,9 @@
 #ifndef MATHS_HPP_
 #define MATHS_HPP_
 
-inline int add(int, int);
+inline long long add(int, int);
 
-inline int mult(int, int);
+inline long long mult(int, int);
 
 
 #endif
